verifier/snp: Verify report against local certificates
Certificates (VCEK or VLEK) stored in the directory
/etc/kbs/snp/ek will be used to validate attestation
reports.

This is a simple implementation. When an attestation request is
received, it reads the certificates from the aforementioned directory
one by one trying to verify the report. At the end, if the report cannot
be verified by a cached certificate, it falls back to downloading a VCEK
from KDS.

Certificate Revocation List (CRL) is not supported.

Signed-off-by: Claudio Carvalho <cclaudio@linux.ibm.com>